+++ internal/ui/model/dialog_actions.go
@@ -598,6 +598,12 @@ func (m *UI) applyChromeDialogAction(action dialog.Action) tea.Cmd {
 	case dialog.ActionOpenDelegationsDashboard:
 		m.dialog.CloseDialog(dialog.CommandsID)
 		cmds = openDelegationsDashboardGuarded(m.com, cmds)
+	case dialog.ActionEnterWorktree:
+		m.dialog.CloseDialog(dialog.CommandsID)
+		cmds = append(cmds, util.CmdHandler(enterWorktreeRequestedMsg{name: msg.Name}))
+	case dialog.ActionExitWorktree:
+		m.dialog.CloseDialog(dialog.CommandsID)
+		cmds = append(cmds, util.CmdHandler(exitWorktreeRequestedMsg{}))
 
 	case dialog.ActionFilePickerSelected:
 		m.dialog.CloseDialog(dialog.FilePickerID)